amazon_scraper/requirements.txt
selenium>=4.15.2
webdriver-manager>=4.0.1
pandas>=2.1.4
requests>=2.31.0
python-dotenv>=1.0.0
